+++ app/webhooks/github/route.ts
@@ -1,17 +1,7 @@
-import { db } from "@/drizzle";
-import { executeStep } from "@giselles-ai/lib/execution";
-import { performFlowExecution } from "@giselles-ai/lib/runner";
-import {
-	createExecutionId,
-	createInitialJobExecutions,
-} from "@giselles-ai/lib/utils";
-import type { Execution, Graph } from "@giselles-ai/types";
 import { Webhooks } from "@octokit/webhooks";
 import type { WebhookEventName } from "@octokit/webhooks-types";
-import { waitUntil } from "@vercel/functions";
 import type { NextRequest } from "next/server";
-import { parseCommand } from "./command";
-import { assertIssueCommentEvent, createOctokit } from "./utils";
+import { WebhookPayloadError, handleEvent } from "./handle_event";
 
 // Extend the max duration of the server actions from this page to 5 minutes
 // https://vercel.com/docs/functions/runtimes#max-duration
@@ -35,120 +25,15 @@ export async function POST(request: NextRequest) {
 	const id = request.headers.get("X-GitHub-Delivery") ?? "";
 	const name = request.headers.get("X-GitHub-Event") as WebhookEventName;
 	const rawPayload = JSON.parse(body);
-	const event = { id, name, payload: rawPayload };
-	assertIssueCommentEvent(event);
-	const payload = event.payload;
 
-	const command = parseCommand(payload.comment.body);
-	if (command === null) {
-		return new Response(
-			`Command not found. payload: ${JSON.stringify(payload)}`,
-			{ status: 400 },
-		);
+	try {
+		await handleEvent({ id, name, payload: rawPayload });
+	} catch (e) {
+		if (e instanceof WebhookPayloadError) {
+			return new Response(e.message, { status: 400 });
+		}
+		throw e;
 	}
-	if (payload.installation === undefined) {
-		return new Response(
-			`Installation not found. payload: ${JSON.stringify(payload)}`,
-			{ status: 400 },
-		);
-	}
-	const octokit = await createOctokit(payload.installation.id);
-
-	const integrationSettings = await db.query.githubIntegrationSettings.findMany(
-		{
-			where: (gitHubIntegrationSettings, { eq, and }) =>
-				and(
-					eq(
-						gitHubIntegrationSettings.repositoryFullName,
-						payload.repository.full_name,
-					),
-					eq(gitHubIntegrationSettings.callSign, command.callSign),
-				),
-		},
-	);
-
-	waitUntil(
-		Promise.all(
-			integrationSettings.map(async (integrationSetting) => {
-				const agent = await db.query.agents.findFirst({
-					where: (agents, { eq }) =>
-						eq(agents.dbId, integrationSetting.agentDbId),
-				});
-				if (agent === undefined || agent.graphUrl === null) {
-					return;
-				}
-				const graph = await fetch(agent.graphUrl).then(
-					(res) => res.json() as unknown as Graph,
-				);
-				const flow = graph.flows.find(
-					(flow) => flow.id === integrationSetting.flowId,
-				);
-				if (flow === undefined) {
-					console.warn(
-						`GitHubIntegrationSetting: ${integrationSetting.id}, flow not found`,
-					);
-					return;
-				}
-				const executionId = createExecutionId();
-				const jobExecutions = createInitialJobExecutions(flow);
-				const flowRunStartedAt = Date.now();
-
-				let initialExecution: Execution = {
-					id: executionId,
-					status: "running",
-					flowId: flow.id,
-					jobExecutions,
-					artifacts: [],
-					runStartedAt: flowRunStartedAt,
-				};
-
-				const overrideData = integrationSetting.eventNodeMappings
-					.map((eventNodeMapping) => {
-						switch (eventNodeMapping.event) {
-							case "comment.body":
-								return {
-									nodeId: eventNodeMapping.nodeId,
-									data: command.content,
-								};
-							case "issue.title":
-								return {
-									nodeId: eventNodeMapping.nodeId,
-									data: payload.issue.title,
-								};
-							default:
-								return null;
-						}
-					})
-					.filter((overrideData) => overrideData !== null);
-				const finalExecution = await performFlowExecution({
-					initialExecution,
-					executeStepFn: (stepId) =>
-						executeStep({
-							agentId: agent.id,
-							flowId: flow.id,
-							executionId: executionId,
-							stepId: stepId,
-							artifacts: initialExecution.artifacts,
-							overrideData,
-						}),
-					onExecutionChange: (execution) => {
-						initialExecution = execution;
-					},
-				});
-
-				await octokit.request(
-					"POST /repos/{owner}/{repo}/issues/{issue_number}/comments",
-					{
-						owner: payload.repository.owner.login,
-						repo: payload.repository.name,
-						issue_number: payload.issue.number,
-						body: finalExecution.artifacts[finalExecution.artifacts.length - 1]
-							.object.content,
-					},
-				);
-			}),
-		),
-	);
 
 	return new Response("Accepted", { status: 202 });
 }
